refactor: renamed input ports to use case and the interector classes to handlers
