+++ src/main/java/com/fiap/grupo186/techchallenge/application/handlers/UpdateStatusOrderHandler.java
@@ -1,7 +1,7 @@
-package com.fiap.grupo186.techchallenge.application.usecases;
+package com.fiap.grupo186.techchallenge.application.handlers;
 
 import com.fiap.grupo186.techchallenge.application.ports.OrderRepositoryPort;
-import com.fiap.grupo186.techchallenge.application.ports.UpdateStatusOrderPort;
+import com.fiap.grupo186.techchallenge.application.ports.UpdateStatusOrderUseCase;
 import com.fiap.grupo186.techchallenge.domains.kitchen.models.OrderStatus;
 import com.fiap.grupo186.techchallenge.domains.kitchen.services.KitchenService;
 import org.springframework.stereotype.Service;
@@ -9,11 +9,11 @@ import org.springframework.stereotype.Service;
 import java.util.UUID;
 
 @Service
-public class UpdateStatusOrderUseCase implements UpdateStatusOrderPort {
+public class UpdateStatusOrderHandler implements UpdateStatusOrderUseCase {
     private final OrderRepositoryPort repository;
     private final KitchenService service;
 
-    public UpdateStatusOrderUseCase(OrderRepositoryPort repository, KitchenService service) {
+    public UpdateStatusOrderHandler(OrderRepositoryPort repository, KitchenService service) {
         this.repository = repository;
         this.service = service;
     }
