+++ src/main/java/com/fiap/grupo186/techchallenge/application/ports/CreatOrderUseCase.java
@@ -3,6 +3,6 @@ package com.fiap.grupo186.techchallenge.application.ports;
 import com.fiap.grupo186.techchallenge.application.dtos.PreOrderDTO;
 import com.fiap.grupo186.techchallenge.domains.kitchen.models.Order;
 
-public interface CreatOrderPort {
+public interface CreatOrderUseCase {
     Order execute(PreOrderDTO dto);
 }
