+++ src/main/java/com/fiap/grupo186/techchallenge/application/ports/UpdateStatusOrderUseCase.java
@@ -4,6 +4,6 @@ import com.fiap.grupo186.techchallenge.domains.kitchen.models.OrderStatus;
 
 import java.util.UUID;
 
-public interface UpdateStatusOrderPort {
+public interface UpdateStatusOrderUseCase {
     void execute(UUID orderId, OrderStatus status);
 }
